byoai-script.py
import os
import yaml

# Updated environment variables
CONTEXT_WORKFLOW_DIR = os.environ.get('CONTEXT_WORKFLOW_DIR', 'workflows')
CONTEXT_AGENT_PORT = os.environ.get('CONTEXT_AGENT_PORT', '8000')
CONTEXT_AGENT_HOST = os.environ.get('CONTEXT_AGENT_HOST', '0.0.0.0')

def load_workflow(workflow_file):
    with open(os.path.join(CONTEXT_WORKFLOW_DIR, workflow_file), 'r') as f:
        return yaml.safe_load(f)

def execute_workflow(workflow):
    # Implement workflow execution logic here
    pass

def main():
    # Main logic for BYOAI
    print(f"BYOAI agent running on {CONTEXT_AGENT_HOST}:{CONTEXT_AGENT_PORT}")
    print(f"Loading workflows from {CONTEXT_WORKFLOW_DIR}")
    
    # Add integration with Swarm's decentralized agent concept
    from swarm_integration import SwarmIntegration
    swarm = SwarmIntegration()
    swarm.start()

if __name__ == "__main__":
    main()
