pyproject.toml
[project]
name = "repl-nix-attachmentanalyzer"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "pyyaml>=6.0.2",
]
